requirements.txt
numpy==1.21.2
Pillow==10.2.0
skimage==0.0
